payment/requirements.txt
bottle==0.13.0
python-dotenv==1.0.1
yookassa==3.0.1
